Add Local permission

+++ backend.py
@@ -1,7 +1,43 @@
-from database_setup import DBSession, Category, Item
+from database_setup import DBSession, Category, Item, User
 from sqlalchemy import desc
+from functools import wraps
+from flask import request, redirect, url_for, session as login_session
 
 session = DBSession()
+
+def login_required(f):
+    @wraps(f)
+    def decorated_function(*args, **kwargs):
+        if login_session.get('logged_in') is None:
+            return redirect(url_for('login', next=request.url))
+        return f(*args, **kwargs)
+    return decorated_function
+
+
+def create_user():
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'], picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
+def isCreator(user_id):
+    creator = getUserInfo(user_id)
+    if creator.id != login_session.get('user_id'):
+        return creator, False
+    return creator, True
 
 def show_catalog():
     return session.query(Category).all()
@@ -15,6 +51,7 @@ def get_category(category):
 def new_category(name):
     category = Category()
     category.name = name
+    category.user_id = login_session['user_id']
     session.add(category)
     session.commit()
     return session.query(Category).filter(Category.name==name).first()
@@ -35,6 +72,7 @@ def new_item(category_name, name, description):
     item.name = name
     item.description = description
     item.category = category
+    item.user_id  = login_session['user_id']
     session.add(item)
     session.commit()
     return session.query(Item).filter(Item.category==category, Item.name==name).first()
@@ -53,7 +91,8 @@ def edit_item(old_category_name, old_item_name, name, description, category_name
 def del_item(category_name, item_name):
     category = session.query(Category).filter(Category.name==category_name).first()
     item = session.query(Item).filter(Item.category==category, Item.name==item_name).first()
-    if item is not None:
+    creator, editable = isCreator(item.user_id)
+    if item is not None and editable:
         session.delete(item)
         session.commit()
         return True
